Relevanssi Version 4.13.3.1

+++ wp-content/plugins/relevanssi/lib/uninstall.php
@@ -50,6 +50,11 @@ function relevanssi_drop_database_tables() {
  * @global object $wpdb The WordPress database interface.
  */
 function relevanssi_uninstall_free() {
+	if ( defined( 'RELEVANSSI_PREMIUM' ) && RELEVANSSI_PREMIUM && ! defined( 'UNINSTALLING_RELEVANSSI_PREMIUM' ) ) {
+		// Relevanssi Premium exists, do not drop the tables.
+		return;
+	}
+
 	delete_option( 'relevanssi_admin_search' );
 	delete_option( 'relevanssi_bg_col' );
 	delete_option( 'relevanssi_cat' );
@@ -70,6 +75,7 @@ function relevanssi_uninstall_free() {
 	delete_option( 'relevanssi_excerpt_type' );
 	delete_option( 'relevanssi_excerpts' );
 	delete_option( 'relevanssi_exclude_posts' );
+	delete_option( 'relevanssi_expand_highlights' );
 	delete_option( 'relevanssi_expand_shortcodes' );
 	delete_option( 'relevanssi_extag' );
 	delete_option( 'relevanssi_fuzzy' );
@@ -99,6 +105,7 @@ function relevanssi_uninstall_free() {
 	delete_option( 'relevanssi_post_type_weights' );
 	delete_option( 'relevanssi_punctuation' );
 	delete_option( 'relevanssi_respect_exclude' );
+	delete_option( 'relevanssi_seo_noindex' );
 	delete_option( 'relevanssi_show_matches' );
 	delete_option( 'relevanssi_show_matches_text' );
 	delete_option( 'relevanssi_show_post_controls' );
@@ -111,7 +118,6 @@ function relevanssi_uninstall_free() {
 	delete_option( 'relevanssi_title_boost' );
 	delete_option( 'relevanssi_trim_logs' );
 	delete_option( 'relevanssi_txt_col' );
-	delete_option( 'relevanssi_word_boundaries' );
 	delete_option( 'relevanssi_wpml_only_current' );
 
 	// Unused options, removed in case they are still left.
@@ -130,6 +136,8 @@ function relevanssi_uninstall_free() {
 	delete_option( 'relevanssi_custom_taxonomies' );
 	delete_option( 'relevanssi_taxonomies_to_index' );
 	delete_option( 'relevanssi_highlight_docs_external' );
+	delete_option( 'relevanssi_word_boundaries' );
+	delete_option( 'relevanssi_expst' );
 
 	global $wpdb;
 	$wpdb->query( "DELETE FROM $wpdb->postmeta WHERE meta_key = '_relevanssi_noindex_reason'" );
